AI_Employee_Vault-silver/scripts/requirements.txt
# AI Employee Watcher Dependencies
# Install with: pip install -r requirements.txt

# =====================
# BRONZE TIER
# =====================

# File system monitoring
watchdog>=4.0.0

# =====================
# SILVER TIER
# =====================

# Gmail Watcher
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0

# LinkedIn/WhatsApp Watcher (Playwright)
playwright>=1.40.0

# PDF Reader
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Excel Reader
openpyxl>=3.1.0
pandas>=2.0.0

# MCP Server
mcp>=0.1.0
httpx>=0.25.0

# =====================
# GENERAL
# =====================

# Environment variables
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0
